feat(runs): Stale state + connect-based liveness probe
Adds a fourth RunStatus variant (Stale) and fixes the long-standing
classifier bug where runs killed by SIGKILL / OOM / crash showed as
Running indefinitely because the v0.8 probe only checked whether
the control-socket FILE existed.

The probe is now a real UnixStream::connect() — dead sockets return
ECONNREFUSED almost instantly, so the classifier costs no more than
before. Stale fires when both:

  1. control socket connect fails, AND
  2. summary.jsonl has not been written within STALENESS_THRESHOLD
     (4h, matching the upcoming `pitboss prune --older-than`
     default).

A pure classify_status() helper splits the four-way decision out of
the file-system caller so every branch is covered by unit tests
without any tempdir setup.

Rendering: e.status.label() returns "stale", which flows
automatically into both `pitboss-tui list` and the TUI run picker
without extra plumbing. Per-state coloring is intentionally deferred
— the new label is the user-visible signal that matters.

Refactor:
  * RunStatus / RunEntry / collect_run_entries / runs_base_dir /
    count_jsonl_tasks / format_mtime moved from pitboss-tui::runs to
    pitboss-cli::runs. Natural home — pitboss-tui already depends on
    pitboss-cli.
  * pitboss-tui::runs is now a thin `pub use pitboss_cli::runs::*;`
    re-export shim so existing TUI imports keep working unchanged.

Unblocks PR 2.B (`pitboss prune`), which will reuse the classifier
and match on Stale by default.
