+++ crates/pitboss-cli/src/lib.rs
@@ -13,6 +13,7 @@ pub mod dispatch;
 pub mod manifest;
 pub mod mcp;
 pub mod notify;
+pub mod runs;
 pub mod shared_store;
 pub mod status;
 pub mod tui_table;
